requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fastapi==0.115.11
filetype==1.2.0
google-ai-generativelanguage==0.6.16
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.69.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.20
langchain-core==0.3.43
langchain-google-genai==2.0.11
langchain-openai==0.3.8
langchain-text-splitters==0.3.6
langsmith==0.3.13
loguru==0.7.3
openai==1.65.4
orjson==3.10.15
packaging==24.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.1
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
zstandard==0.23.0
